LayoutTests/platform/ios-simulator/editing/deleting/delete-select-all-001-expected.txt

One	Two	Three
One	Two	Three
One	Two	Three
execDeleteCommand: <table border="1"> <tbody><tr><td id="test">One</td><td>Two</td><td>Three</td></tr> <tr><td>One</td><td>Two</td><td>Three</td></tr> <tr><td>One</td><td>Two</td><td>Three</td></tr> </tbody></table>
